+++ cscglobal-caplugin/Constants.cs
@@ -13,7 +13,6 @@ public class Constants
     public static string CscGlobalApiKey = "ApiKey";
     public static string BearerToken = "BearerToken";
     public static string DefaultPageSize = "DefaultPageSize";
-    public static string TemplateSync = "TemplateSync";
     public static string SyncFilterDays = "SyncFilterDays";
 }
     
@@ -21,13 +20,16 @@ public class ProductIDs
 {
     public static List<String> productIds = new List<string>()
     {
-        "CSC TrustedSecure Premium Certificate",
-        "CSC TrustedSecure EV Certificate",
-        "CSC TrustedSecure UC Certificate",
-        "CSC TrustedSecure Premium Wildcard Certificate",
-        "CSC TrustedSecure Domain Validated SSL",
-        "CSC TrustedSecure Domain Validated Wildcard SSL",
-        "CSC TrustedSecure Domain Validated UC Certificate"
+        "CSC TrustedSecure OV",
+        "CSC TrustedSecure OV Wildcard",
+        "CSC TrustedSecure OV, Multiple Names",
+        "CSC TrustedSecure EV",
+        "CSC TrustedSecure DV",
+        "CSC TrustedSecure DV Wildcard",
+        "CSC TrustedSecure DV, Multiple Names",
+        "CSC TrustedSecure EV, Multiple Names",
+        "CSC TrustedSecure OV Wildcard, Multiple Names",
+        "CSC TrustedSecure DV Wildcard, Multiple Names"
     };
 }